clean up and move index tests to extended test
